backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.27
pydantic==2.5.2
requests==2.31.0
python-dotenv==1.0.0
httpx==0.25.2
python-multipart==0.0.7
aiohttp==3.9.3 